+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20190502095422.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\DBAL\Schema\Schema;
+
+/**
+ * Auto-generated Migration: Evaluate and comment a contact
+ */
+class Version20190502095422 extends AbstractMigration
+{
+    /**
+     * @param Schema $schema
+     */
+    public function up(Schema $schema)
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE contact ADD evaluation SMALLINT NOT NULL, ADD comment LONGTEXT NOT NULL');
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    public function down(Schema $schema)
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE contact DROP evaluation, DROP comment');
+    }
+}